removing template references first step
